+++ apps/web/src/queries/cars/makes/index.ts
@@ -3,3 +3,4 @@ export * from "./entity-breakdowns";
 export * from "./entity-checks";
 export * from "./get-make-from-slug";
 export * from "./grouped-makes";
+export * from "./registration-stats";